feat: add genre and serie entities, repositories, and service interfaces; update category schemas to support nullable fields

+++ src/features/genre/application/use-cases/genre-service.ts
@@ -2,36 +2,32 @@ import { DI_TYPES } from '@src/core/di/types';
 import { AppError } from '@src/core/errors/custom-error';
 import { inject, injectable } from 'inversify';
 import {
-  CategoryCreateDto,
-  CategoryDetailDto,
-  CategoryUpdateDto,
-  Id,
-} from '../dto/category-dtos';
-import { CategoryEntity } from '../../domain/entities/category';
-import { CategoryRepository } from '../../domain/repository/category-repository';
-import { ICategoryService } from './interfaces/category-service-interface';
+  GenreCreateDto,
+  GenreDetailDto,
+  GenreUpdateDto,
+} from '../dto/genre-dtos';
 import logger from '@src/core/utils/logger/logger';
+import { GenreRepository } from '../../domain/repository/genre-repository';
+import { IGenreService } from './interfaces/genre-service-interface';
+import { GenreEntity } from '../../domain/entities/genre-entity';
+import { Id } from '@src/core/types';
 
 @injectable()
-export class CategoryService implements ICategoryService {
-  private readonly repository: CategoryRepository;
-  constructor(
-    @inject(DI_TYPES.CategoryRepository) repository: CategoryRepository,
-  ) {
+export class GenreService implements IGenreService {
+  private readonly repository: GenreRepository;
+  constructor(@inject(DI_TYPES.GenreRepository) repository: GenreRepository) {
     this.repository = repository;
   }
-  async create(data: CategoryCreateDto): Promise<CategoryDetailDto> {
+  async create(data: GenreCreateDto): Promise<GenreDetailDto> {
     try {
       const existed = await this.repository.getByNameAsync(data.name);
       if (existed) {
-        throw AppError.forbidden('existed category.');
+        throw AppError.forbidden('existed Genre.');
       }
 
-      const input: Partial<CategoryEntity> = {
+      const input: Partial<GenreEntity> = {
         name: data.name,
-        description: data.description ?? '',
-        cover: data.cover ?? '',
-        updatedAt: new Date(),
+        description: data.description ?? undefined,
       };
 
       const res = await this.repository.create(input);
@@ -41,27 +37,31 @@ export class CategoryService implements ICategoryService {
       throw new Error(`error: ${error}`);
     }
   }
-  async update(id: Id, data: CategoryUpdateDto): Promise<string> {
+  async update(id: Id, data: GenreUpdateDto): Promise<string> {
     try {
       const existed = await this.repository.getById(id);
       if (!existed) {
-        throw AppError.notFound('category not found.');
+        throw AppError.notFound('Genre not found.');
       }
 
-      await this.repository.update(id, {
+      const updatedDate: Partial<GenreEntity> = {
         ...existed,
         name: data.name ?? existed.name,
-        description: data.description ?? '',
-        cover: data.cover ?? '',
+        description:
+          data.description === undefined
+            ? existed.description
+            : (data.description ?? ''),
         updatedAt: new Date(),
-      });
+      };
+
+      await this.repository.update(id, updatedDate);
       return existed.id;
     } catch (error) {
       logger.error(error);
       throw error;
     }
   }
-  async getList(): Promise<Array<CategoryDetailDto>> {
+  async getList(): Promise<Array<GenreDetailDto>> {
     try {
       const res = await this.repository.getList();
 
@@ -71,7 +71,7 @@ export class CategoryService implements ICategoryService {
       throw AppError.internalServer('Internal server error.');
     }
   }
-  async getById(id: string): Promise<CategoryDetailDto | null> {
+  async getById(id: string): Promise<GenreDetailDto | null> {
     try {
       const res = await this.repository.getById(id);
 
@@ -87,20 +87,23 @@ export class CategoryService implements ICategoryService {
   }
   async delete(id: string): Promise<string> {
     try {
+      const existed = await this.repository.getById(id);
+      if (!existed) {
+        throw AppError.notFound('Genre not found.');
+      }
+
       await this.repository.delete(id);
       return id;
     } catch (error) {
-      logger.error(error);
       throw AppError.internalServer('Internal server error.');
     }
   }
 
-  private _convertToResultDto(entity: CategoryEntity): CategoryDetailDto {
+  private _convertToResultDto(entity: GenreEntity): GenreDetailDto {
     return {
       id: entity.id,
       name: entity.name,
       description: entity.description,
-      cover: entity.cover,
       updatedAt: entity.updatedAt,
     };
   }
